[#102] Refactor ConfigMenu
